updated delete category and tag

+++ app/Http/Controllers/TagsController.php
@@ -96,6 +96,10 @@ class TagsController extends Controller
     public function destroy(Tag $tag)
     {
         //
+        if ($tag->posts->count() > 0) {
+            session()->flash('error', 'Tags cannot be delete because it has some posts');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success', 'Category Delete successfully');
         return redirect(route('tags.index'));
